pyproject.toml
[tool.poetry]
name = "dabini-engine"
version = "0.1.0"
description = "A FastAPI application using LangChain and OpenAI"
authors = ["Choi Yeonung <yonsweng@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
cloud-sql-python-connector = "^1.17.0"
fastapi = "^0.115.11"
langchain-community = "^0.3.19"
langchain-core = "^0.3.45"
langchain-openai = "^0.3.7"
pymysql = "^1.1.1"
python = "^3.13"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.39"
uvicorn = "^0.34.0"

[tool.poetry.scripts]
dev = "dabini_engine.main:dev"
start = "dabini_engine.main:main"

[build-system]
requires = ["poetry-core>=2.1.1"]
build-backend = "poetry.core.masonry.api"
